[ExchangeSimulator] Add spot, margin and future
