+++ octobot_trading/exchanges/types/future_exchange.py
@@ -14,9 +14,10 @@
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
 from octobot_trading.enums import PositionSide
+from octobot_trading.exchanges.abstract_exchange import AbstractExchange
 
 
-class FutureExchange:
+class FutureExchange(AbstractExchange):
     LONG_STR = "long"
     SHORT_STR = "short"
