+++ octobot_trading/exchanges/types/margin_exchange.py
@@ -13,7 +13,8 @@
 #
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
+from octobot_trading.exchanges.abstract_exchange import AbstractExchange
 
 
-class MarginExchange:
+class MarginExchange(AbstractExchange):
     pass
